fix: Remove duplicate `create_internal_customer` util
- Introduced via https://github.com/frappe/erpnext/pull/27086
