+++ erpnext/stock/doctype/delivery_note/test_delivery_note.py
@@ -435,7 +435,8 @@ class TestDeliveryNote(unittest.TestCase):
 		company = frappe.db.get_value('Warehouse', 'Stores - TCP1', 'company')
 		customer = create_internal_customer(
 			customer_name="_Test Internal Customer 2",
-			allowed_company="_Test Company with perpetual inventory"
+			represents_company="_Test Company with perpetual inventory",
+			allowed_to_interact_with="_Test Company with perpetual inventory"
 		)
 
 		set_valuation_method("_Test Item", "FIFO")
